tools: Move numpy-related utils to separate module
