+++ devito/tools/__init__.py
@@ -1,5 +1,5 @@
 from devito.tools.utils import *  # noqa
-
+from devito.tools.numpy_helper import *  # noqa
 from devito.tools.abc import *  # noqa
 from devito.tools.algorithms import *  # noqa
 from devito.tools.data_structures import *  # noqa
