+++ devito/tools/numpy_helper.py
@@ -1,0 +1,20 @@
+import ctypes
+import numpy as np
+
+__all__ = ['numpy_to_ctypes', 'numpy_to_mpitypes']
+
+
+def numpy_to_ctypes(dtype):
+    """Map numpy types to ctypes types."""
+    return {np.int32: ctypes.c_int,
+            np.float32: ctypes.c_float,
+            np.int64: ctypes.c_int64,
+            np.float64: ctypes.c_double}[dtype]
+
+
+def numpy_to_mpitypes(dtype):
+    """Map numpy types to MPI datatypes."""
+    return {np.int32: 'MPI_INT',
+            np.float32: 'MPI_FLOAT',
+            np.int64: 'MPI_LONG',
+            np.float64: 'MPI_DOUBLE'}[dtype]
