+++ devito/tools/utils.py
@@ -7,10 +7,9 @@ from operator import attrgetter, mul
 import numpy as np
 import sympy
 
-__all__ = ['prod', 'as_tuple', 'is_integer', 'generator', 'grouper', 'split',
-           'roundm', 'powerset', 'invert', 'flatten', 'single_or', 'filter_ordered',
-           'filter_sorted', 'numpy_to_ctypes', 'numpy_to_mpitypes', 'ctypes_to_C',
-           'ctypes_pointer', 'pprint', 'sweep', 'as_mapper']
+__all__ = ['prod', 'as_tuple', 'is_integer', 'generator', 'grouper', 'split', 'roundm',
+           'powerset', 'invert', 'flatten', 'single_or', 'filter_ordered', 'as_mapper',
+           'filter_sorted', 'ctypes_to_C', 'ctypes_pointer', 'pprint', 'sweep']
 
 
 def prod(iterable):
@@ -148,22 +147,6 @@ def filter_sorted(elements, key=None):
     if key is None:
         key = attrgetter('name')
     return sorted(filter_ordered(elements, key=key), key=key)
-
-
-def numpy_to_ctypes(dtype):
-    """Map numpy types to ctypes types."""
-    return {np.int32: ctypes.c_int,
-            np.float32: ctypes.c_float,
-            np.int64: ctypes.c_int64,
-            np.float64: ctypes.c_double}[dtype]
-
-
-def numpy_to_mpitypes(dtype):
-    """Map numpy types to MPI datatypes."""
-    return {np.int32: 'MPI_INT',
-            np.float32: 'MPI_FLOAT',
-            np.int64: 'MPI_LONG',
-            np.float64: 'MPI_DOUBLE'}[dtype]
 
 
 def ctypes_to_C(ctype):
